tcg: add sync_temp opcode
Currently every field of CPUArchState can be accessed from the TCG-generated code
as a memory location or as a global but not both. In order to be able to mix
these two approaches we need to restore consistency between value of global
(possibly kept on register) and value in corresponding memory location.

Introduce sync_temp TCGOpcode which instructs register allocator to
save value of a global into its memory location.

Signed-off-by: Kirill Batuzov <batuzovk@ispras.ru>